backend/requirements.txt
flask==3.0.2
flask-cors==4.0.0
python-dotenv==1.0.0
azure-ai-documentintelligence==1.0.0
azure-identity==1.15.0
python-multipart==0.0.6
ruff==0.11.2